requirements.txt
# Core ML
torch>=2.2.0
torch-geometric>=2.5.0
transformers>=4.40.0
sentence-transformers>=3.0.0
scikit-learn>=1.5.0
numpy>=1.26.0
pandas>=2.2.0

# Vector Store
faiss-cpu>=1.7.0

# Streaming
kafka-python>=2.0.2

# Feature Store
feast>=0.38.0

# API
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1

# Evaluation
mlflow>=2.13.0

# LLM
openai>=1.30.0

# Testing
pytest>=8.2.0
pytest-cov>=5.0.0
httpx>=0.27.0  # FastAPI test client
